requirements.txt
streamlit>=1.42.0
Authlib>=1.3.2
supabase
python-dotenv
streamlit-oauth